Feature/talos-cluster (#1476)
* Add resource requests and limits to gateway configuration files

This commit updates the `gateway-http2-values.yaml` and `gateway-values.yaml` files to include resource requests and limits for CPU and memory. The changes specify a request of 100m CPU and 512Mi memory, while the limits for CPU remain unset and memory is capped at 512Mi. These adjustments enhance resource management and ensure better performance in Kubernetes deployments.

* Update resource requests and limits in kiali values.yaml

This commit modifies the resource requests and limits for CPU and memory in the kiali values.yaml file. The CPU request is increased to 365m, while the memory request and limit are both set to 300Mi. These changes aim to optimize resource allocation for better performance in Kubernetes deployments.

* Update resource requests and limits in Loki, Mimir, and Tempo values files

This commit modifies the resource requests and limits for CPU and memory across the Loki, Mimir, and Tempo configuration files. Key changes include reducing memory requests and limits for various caches in Loki and Mimir, and adding resource specifications for multiple components in Tempo. These adjustments aim to optimize resource allocation and improve performance in Kubernetes deployments.